Aggregation split into test and header file
